% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{add_diam_on_screen}
\alias{add_diam_on_screen}
\title{Calculates screen diameter for each model.}
\usage{
add_diam_on_screen(x, alpha_col, screen_distance)
}
\description{
x = model dataframe containing alpha angles data
alpha_col = column identifier for alpha angles data
screen_distance = distance simulation will be viewed at
}
\details{
A new column 'diam_on_screen' will be added to the dataframe
}
\keyword{internal}
