% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{add_screen_diam}
\alias{add_screen_diam}
\title{Calculates screen diameter for a model based on alpha and a viewing
distance}
\usage{
add_screen_diam(x, alpha = NULL, screen_distance)
}
\arguments{
\item{x}{numeric. Output of \code{attack_model} or \code{attack_model_whale} function,
or other object containing \code{alpha} angles data.}

\item{screen_distance}{numeric. distance from screen in same units as those
used in model}

\item{alpha_col}{numeric. The column of alpha values. Automatically detected
in most models, but can be specified here.}
}
\value{
Original object or data frame with 'screen_diam' column added, in
same units as used in model and screen_distance. Rounded to 2 decimals
places. Because of values approaching infinty at very close distances, any
value over 1000 is replaced with 1000, so use appropriate units (cm is
best).
}
\description{
Calculates screen diameter for a model based on alpha and a
viewing distance
}
\keyword{internal}
