% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{af_speeds_mont.rd}
\alias{af_speeds_mont.rd}
\title{Speeds for a humpback whale feeding on anchovies in Monterey Bay}
\format{A numeric vector consisting of 688 values of speed in cm/s.}
\usage{
af_speeds_mont.rd
}
\description{
Multiple measurements of speed (cm/s) of a humpback whale feeding on anchovies in Monterey Bay.
}
\author{
Nicholas Carey, Dave Cade
}
\keyword{datasets}
