% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{af_speeds_socal.rd}
\alias{af_speeds_socal.rd}
\title{Speeds for a humpback whale feeding on anchovies in Southern California}
\format{A numeric vector consisting of 1501 values of speed in cm/s.}
\usage{
af_speeds_socal.rd
}
\description{
Multiple measurements of speed (cm/s) of a humpback whale feeding on anchovies in Southern California.
}
\author{
Nicholas Carey, Dave Cade
}
\keyword{datasets}
