% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack_model.R
\name{attack_model}
\alias{attack_model}
\title{Attack Model}
\usage{
attack_model(speed, frequency = 60, body_length = 10.5, model_length =
 NULL, simple_output = TRUE, cALT = c(0.89, 1.66, 2.06), plot = TRUE,
 plot_from = 0, plot_to = NULL)
}
\arguments{
\item{speed}{numeric. A vector of speeds of the modelled approaching attacker.
Must be at same frequency in Hz as `frequency`. If a data.frame is entered
the first colummn is used. A single constant speed value can also be used;
the function will repeat this the required number of times at the correct
frequency based on `model_length`.}

\item{model_length}{integer. Total length of the model. Required if `speed` is
a single value, when along with `frequency` it determines the distance the
predator starts at. If `speed` is a vector this can be left NULL, in which
case it is assumed the predator reaches the prey on the last value, and the
length of the `speed` vector determines total length of model.
Alternatively, `model_length` can be used to set a different capture point
along the speed vector, in which case its value must be less than the total
length of `speed`.}

\item{frequency}{numeric. Frequency (Hz) of the model (frame rate of any
resulting animation). Must be same frequency in Hz as \code{speed}}

\item{simple_output}{logical. Choose structure of output. If TRUE, a simple
dataframe of the model is returned, otherwise output is a List object
containing the final model and input parameters.}

\item{plot}{logical. Choose to plot result.}

\item{plot_from}{numeric. Time on x-axis to plot from.}

\item{plot_to}{numeric.  Time on x-axis to plot to.}

\item{attacker_length}{numeric. Length of the attacker. Must be same units as
maximum width.}

\item{alpha}{numeric. Vector of three values of critical ALT range: lower,}

\item{dadt}{numeric. Vector of three values of critical ALT range: lower,
mean, upper. Enter as c(1,2,3). Helps put range on plot. Mean is not plotted
for now. If any are not reached in the scenario there should be a message.
If upper range is not reached, range is plotted from lower value to end of
model, i.e. `model_length` location.}
}
\description{
\code{attack_model} models the visual aspects of an attack by a
 predator on a prey

 Models an attack by a predator on a prey using parameters such as size,
 speed, shape etc. From the prey perspective, it calculates the visual angle
 of the diameter of the attacker (alpha, or \strong{a}) in radians, and the
 rate of change of this angle (\strong{da/dt} in radians/s).

 Input units: `frequency` is in \code{Hz}. If `speed` is a vector it should
 be at the same frequency. Body measurements and `speed` units can be any
 unit, but should be consistent. i.e. all in `cm`, speeds in `cm/s`, etc.

 `speed` can be either a constant value or vector of speeds (at the
 correct`frequency`). If a single constant value, a `model_length` is
 required in order to calculate the distance the predator starts the attack
 from. For example, at a `frequency` of 60 Hz and `model_length` of 180, the
 model will be three seconds in duration. At a constant `speed` of 600 cm/s,
 the predator will start from approximately 1800 cm away.

 The `model_length` operator can also be used to vary where on variable speed
 vectors the predator times reaching the prey, to examine how this affects
 perceived alphas. If left NULL, with speed vectors by default the model ends
 with the predator reaching the prey on the last speed value, however
 `model_length` can be used to set this to be earlier.

 Models end when the nose (most anterior part in the profiles) of the
 predator reaches the prey (last row of the data.frame, where time and
 distance equal zero).
}
\details{
In order to correctly model the widest apparent part of the predator
 it requires several morphometric inputs.
}
\examples{
attack_model(...)

}
\author{
Nicholas Carey - \link{nicholascarey@gmail.com}, Dave Cade
 \link{davecade@stanford.edu},
}
