% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bw_speeds.rd}
\alias{bw_speeds.rd}
\title{Speeds for blue whale feeding on krill}
\format{A numeric vector consisting of 1501 values of speed in cm/s.}
\usage{
bw_speeds.rd
}
\description{
Multiple measurements of speed (cm/s) of a blue whale feeding on krill.
}
\author{
Nicholas Carey, Dave Cade
}
\keyword{datasets}
