% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{calc_alpha}
\alias{calc_alpha}
\title{Calculate alpha in radians}
\usage{
calc_alpha(diameter, distance)
}
\description{
Note - only works for complete objects with same diameter and distance. So
can be used for alpha of max_girth, but not of jaws.
}
\keyword{internal}
