% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{calc_dadt}
\alias{calc_dadt}
\title{Calculate da/dt in radians}
\usage{
calc_dadt(speed, diameter, distance)
}
\description{
Note - only works for complete objects with same diameter and distance. So
can be used for dadt of max_girth, but not of jaws.
}
\keyword{internal}
