% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{closest}
\alias{closest}
\title{Get closest matching value}
\usage{
closest(this, there)
}
\arguments{
\item{this}{numeric. Vector or value. The value(s) to match.}

\item{there}{numeric. The vector in which to find closest matching value(s).}
}
\value{
Index of where closest matching value(s) occur in the target vector.
}
\description{
\code{closest} - takes a numeric vector or single value and
finds the index of closest matching value(s) in a different vector.
}
\details{
Finds the index (i.e. position(s)) of the closest
matching value(s) of the input(s) in a vector
}
\seealso{
\code{\link{closest}}
}
\keyword{internal}
