% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{first_closest}
\alias{first_closest}
\title{Get first AND closest matching value}
\usage{
first_closest(this, there)
}
\arguments{
\item{this}{numeric. Vector or value. The value(s) to match.}

\item{there}{numeric. The vector in which to find first closest matching value(s).}
}
\value{
Index of where first_closest matching value(s) occur in the target vector.
}
\description{
\code{first_closest} - takes a single value and finds the index
of first time that value is exceeded in a vector, then if that value or the
one preceding it is actually closer.
}
\seealso{
\code{\link{first_closest}}
}
\keyword{internal}
