% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{first_over}
\alias{first_over}
\title{Get first close matching value in a generally increasing vector}
\usage{
first_over(this, there)
}
\arguments{
\item{this}{numeric. Single value to find first occurence of.}

\item{there}{numeric. The vector in which to find value.}
}
\value{
Index vector of where the value occurs or is exceeded in the target
  vector.
}
\description{
\code{first_over} - takes a single value and finds the index of
  first occurrence or first time that value is exceeded in a vector.
}
\details{
Finds the index (i.e. position) of the first occurrence or first
  time a value is exceeded in a vector.
}
\keyword{internal}
