% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{get_alpha}
\alias{get_alpha}
\title{extract attacker alpha when a specific da/dt value is exceeded}
\usage{
get_alpha(mod, dAdt)
}
\description{
extract attacker alpha when a specific da/dt value is exceeded
}
\keyword{internal}
