% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{get_dist}
\alias{get_dist}
\title{Extract attacker distance when a specific da/dt value is exceeded}
\usage{
get_dist(mod, ALT)
}
\description{
Extract attacker distance when a specific da/dt value is exceeded
}
\keyword{internal}
