% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{interpolate_widths}
\alias{interpolate_widths}
\title{Interpolates a body profile to the units used in body length}
\usage{
interpolate_widths(profile, max_width_loc, body_length, body_width)
}
\description{
Interpolates a body profile to the units used in body length
}
\details{
Interpolates linearly between each proportional width.
E.g. if 10 prop widtsh are given in the profile and the body length is 100cm,
result gives 101 proportional values from 0 to 100.
}
\keyword{internal}
