% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kf_speeds.rd}
\alias{kf_speeds.rd}
\title{Speeds for a humpback whale feeding on krill in Monterey Bay}
\format{A numeric vector consisting of 649 values of speed in cm/s.}
\usage{
kf_speeds.rd
}
\description{
Multiple measurements of speed (cm/s) of a humpback whale feeding on krill in Monterey Bay.
}
\author{
Nicholas Carey, Dave Cade
}
\keyword{datasets}
