% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack_model.R
\name{attack_model}
\alias{attack_model}
\title{Attack Model}
\usage{
attack_model(speed, model_length = NULL, frequency = 60, body_length =
 NULL, body_width_v = NULL, body_width_h = NULL, profile_v = NULL, profile_h
 = NULL, max_width_loc_v = NULL, max_width_loc_h = NULL, width_filter =
 "mid", simple_output = TRUE, plot = TRUE, plot_from = 0, plot_to = NULL,
 alpha_range = NULL, dadt_range = NULL)
}
\arguments{
\item{speed}{numeric. Either a single constant speed value or vector of speeds
at the same frequency in Hz as \emph{\code{frequency}}. Must be same unit as
\emph{\code{body_length}} per second. If a data.frame is entered the first
colummn is used. For a constant speed value the function will repeat this
the required number of times at the correct frequency based on
\emph{\code{model_length}}.}

\item{model_length}{integer. Total length of the model in rows. Required if
\emph{\code{speed}} is a single value, in which case along with frequency it
determines the distance the predator starts at. If \emph{\code{speed}} is a
vector \emph{\code{model_length}} can be left NULL, in which case it is assumed
the predator reaches the prey on the last value, and the length of the speed
vector determines total length of model. Alternatively,
\emph{\code{model_length}} can be used to set a different capture point along the
speed vector, in which case its value must be less than the total length of
\emph{\code{speed}}.}

\item{frequency}{numeric. Frequency (Hz) of the model, i.e. how many speed and
other measurements per second. Must be same frequency in Hz as
\emph{\code{speed}}.}

\item{body_length}{numeric. Length of the attacker. Must be same units as
\emph{\code{body_width_v}} and \emph{\code{body_width_h}}, and that used in
\emph{\code{speed}}.}

\item{body_width_v}{numeric. Maximum width of the attacker in the vertical
plane.}

\item{body_width_h}{numeric. Maximum width of the attacker in the horizontal
plane.}

\item{profile_v}{numeric. A vector describing the shape of the attacker in the
vertical plane. See details.}

\item{profile_h}{numeric. A vector describing the shape of the attacker in the
horizontal plane. See details.}

\item{max_width_loc_v}{numeric. Location of the maximum girth in the vertical
plane of the predator along the body, if not provided as part of the body
profile inputs. See details.}

\item{max_width_loc_h}{numeric. Location of the maximum girth in the
horizontal plane of the predator along the body, if not provided as part of
the body profile inputs. See details.}

\item{width_filter}{string. Filters apparent widths between vertical and
horizontal planes for each row of the model in various ways. See details.}

\item{simple_output}{logical. Choose structure of output. If TRUE, a simple
data frame of the model is returned, otherwise output is a \emph{\code{list}}
object given an \emph{\code{attack_model}} class, and containing the final model,
input parameters, subset regions, and more.}

\item{plot}{logical. Choose to plot result.}

\item{plot_from}{numeric. Time on x-axis to plot from.}

\item{plot_to}{numeric.  Time on x-axis to plot to.}

\item{alpha_range}{numeric. Vector of two values of alpha. Optional. These
will appear on any plot as a blue region, and if \emph{\code{simple_output =
 FALSE}}, this region of the model is subset out to a separate entry in the
saved \emph{\code{list}} object. If any are not reached in the scenario there
should be a message. If upper range is not reached, it is plotted from lower
value to end of model, i.e. \emph{\code{model_length}} location.}

\item{dadt_range}{numeric. Vector of two values of alpha. Optional. These will
appear on any plot as a green region, and if \emph{\code{simple_output = FALSE}},
this region of the model is subset out to a separate entry in the saved
\emph{\code{list}} object. If any are not reached in the scenario there should be
a message. If upper range is not reached, it is plotted from lower value to
end of model, i.e. \emph{\code{model_length}} location.}
}
\description{
\emph{\code{attack_model}} models the visual aspects of an attack by a
predator on a prey

Using parameters such as predator size, speed, shape etc. this function
models the visual aspects of an attack by the predator on a prey. From the
prey's perspective it calculates the visual angle of the attacker
(\strong{alpha}, or \strong{{α}}) in radians, and the rate of change of this angle
(\strong{{dα/dt}} in radians/s), as well as distance and time to capture.
}
\details{
Information on inputs:
}
\section{Input units}{
 \emph{\code{speed}} should be in the same units as body
measurements (length, width etc.) \emph{per second}, e.g. cm/s. It can be either
a single, constant speed value or a vector of speeds (at the correct
\emph{\code{frequency}}). If a single value, \emph{\code{model_length}} is required in
order to calculate the distance the predator starts the attack from. For
example, at a frequency of 60 Hz and \emph{\code{model_length}} of 180, the model
will be three seconds in duration. Therefore, at a constant speed of 600
cm/s, the predator will start from approximately 1800 cm away.

\emph{\code{frequency}} is in \emph{\code{Hz}}. If \emph{\code{speed}} is a vector it
should be at the same frequency.

\emph{\code{model_length}} is required when \emph{\code{speed}} is a single value,
where it determines the total length of the model. It is also used to vary
where on variable speed vectors the predator times reaching the prey, for
example cooincident with its maximum speed, to examine how this affects
factors such as \strong{{dα/dt}}. If left as the default \emph{\code{model_length =
 NULL}}, the model ends with the predator reaching the prey on the last speed
value.
}

\section{Morphology}{
 In order to correctly model the widest apparent part of
the predator, and thus \strong{{α}}, the function requires several morphometric
inputs:
}

\section{\emph{Length and Width}}{
 \emph{\code{body_length}} is simply the total length
from nose to tail (or equivalent). \emph{\code{body_width_v}} and
\emph{\code{body_width_h}} are the \strong{maximum} body widths in the vertical (i.e.
dorsal:ventral) and horizontal (left:right) planes of the body. These must
all be in the same units.
}

\section{Body Profiles}{
 \emph{\code{profile_v}} and \emph{\code{profile_h}} are vectors
of widths of the predator's body as a proportion of the maximum width going
from the anterior (nose) to the posterior (tail). Therefore, they should
generally start and end on zero (though they don't have to), and all values
must be between zero and 1. They must be regularly spaced, e.g. every 10\%
along the body. They do not have to be the same length (i.e. resolution),
though it is recommended they are. The longer (i.e. higher resolution) these
measurements are, the better their representation of the morphology of the
animal. Actual values of width along the body are calculated to the
resolution of the entered unit of \emph{\code{body_length}} by linearly
interpolating between each proportional width. For example, if the predator
\emph{\code{body_length}} is 1000cm, a width value in cm is calculated at every
cm along the body by interpolating between the proportional widths.
Therefore, use units which will give an appropriate resolution, at least
greater than the profile vector lengths. The higher the numeric value of the
\emph{\code{body_length}} the better, at least three digits is recommended.

You do not need to enter both \emph{\code{profile_v}} and \emph{\code{profile_h}},
only one is required; if one is left NULL the model will use the other to
calculate \strong{{α}}. This can be useful if your predator is always wider in
one plane than the other.

The function calculates the widest apparent width of any part of the
predator's body from the prey's perspective at each iteration of the model,
in both planes. This is \emph{usually} the maximum width of the predator, but it
also depends on the predator's shape. At close distances, more anterior
parts of the predator will appear to be wider, and so result in a higher
\strong{{α}} value.

Note: no section of the body posterior of the maximum width can ever show a
larger apparent visual \strong{{α}} than the maximum width. Therefore, the
function only really needs profile data from the maximum width forward.
However, it is important for the function to work correctly that a full,
evenly spaced profile is provided. Having said that, the proportional width
values posterior of the max width can be placeholder values and it will not
affect calculated \strong{{α}}, as long as they are less than 1 and spaced
correctly.
}

\section{Filtering of maximum apparent width between body planes}{
 Apparent
widths are calculated for both body planes, if two are entered. Which width
at each body segment (e.g. horizontal or vertical) used to calculate \strong{{α}}
is determined via the \emph{\code{width_filter}} operator. This can be the
midpoint value between them (\emph{\code{"mid"}}, the default), maximum value
(\emph{\code{"max"}}), or minimum value (\emph{\code{"min"}}). You can also choose to
use only the vertical or horizontal profile widths exclusively
(\emph{\code{"v"}}, \emph{\code{"h"}}). You can also choose to use the predator's
maximum width in either plane (\emph{\code{max_width_v}}, \emph{\code{max_width_h}})
to calculate \strong{{α}}, in which case all other segments of the body are
ignored, and only \strong{{α}} of the maximum width is determined.

\emph{\code{max_width_loc_v}} and \emph{\code{max_width_loc_h}} are the locations of
the maximum widths of the predator occur along the body as a proportion of
the total body length going from the anterior. They are only necessary if
they are not specified as one of the proportional widths in
\emph{\code{profile_v}} and \emph{\code{profile_h}}, that is none of these have the
value of exactly 1. They can also occur at an intermediate section of the
body, not included as part of the body profiles.
}

\section{Output}{
 By default \emph{\code{simple_output = TRUE}} the output is a data
frame with every row representing a single instance (\emph{\code{frame}}) of the
model at the set \emph{\code{frequency}}. The data frame contains columns for
frame, speed, time, time reversed, distance of the nose of the predator,
\strong{{α}}, and \strong{{dα/dt}}. If \emph{\code{simple_output = FALSE}} the output is a
\emph{\code{list()}} object given class \emph{\code{attack_model}} containing the
above data frame, plus inputs, subset regions (see next section), and data
detailing how each alpha was calculated (interpolated profiles, locations of
maximum apparent width at each frame, etc.). This option greatly increases
the time the function takes to run.
}

\section{Subset regions}{
 \emph{\code{alpha_range}} and \emph{\code{dadt_range}} allow
regions bounded by values of \strong{{α}} and \strong{{dα/dt}} to be subset. Both
inputs are a vector of two numeric values indicating the lower and upper
range of the desired \strong{{α}} and \strong{{dα/dt}} region. The function identifies
the closest matching, first instance of these values in the \emph{\code{$alpha}}
and \emph{\code{dadt}} columns of the model data frame. If \emph{\code{simple_output =
 FALSE}}, these can be found as their own elements in the output
\emph{\code{list()}} object. If \emph{\code{plot = TRUE}} these are also plotted as
blue and green shaded regions, respectively.
}

\section{Plotting}{
 If \emph{\code{plot = TRUE}} a plot is produced showing \emph{speed},
\strong{{α}} and \strong{{dα/dt}}. The X range of the plot can set with
\emph{\code{plot_from}} and \emph{\code{plot_to}}. Optional \emph{\code{alpha_range}} and
\emph{\code{dadt_range}} can also be plotted (see above).
}

\section{General}{
 Models end when the nose (most anterior part in the
profiles) of the predator reaches the prey (last row of the data.frame,
where time and distance equal zero).

At some point, the maximum girth of the predator will not make up the widest
apparent visual angle of the predator, but more anterior segments will
appear to the prey to be wider, and have a higher \strong{{α}}. After
interpolation of body profiles, filtering of apparent widths (see above),
and identification of maximum apparent width at each iteration of the model,
these final widths and their relative distances from the observing prey are
used to calculate the viewing angle, \strong{{α}}, and used to calculate the rate
of change in \strong{{α}}, \strong{{dα/dt}} in radians/s.
}

\author{
Nicholas Carey - \email{nicholascarey@gmail.com}, Dave Cade
\email{davecade@stanford.edu},
}
