% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_dist}
\alias{get_dist}
\title{Extract attacker distance when a specific \strong{{α}} or \strong{{dα/dt}} value
is exceeded}
\usage{
get_dist(mod, dadt, alpha)
}
\arguments{
\item{mod}{list or data.frame. Object of class \code{attack_model},
\code{attack_model_whale}, or \code{data.frame} exported from either of
those functions}

\item{dadt}{numeric. dadt value to find distance exceeded at.}

\item{alpha}{numeric. alpha value to find distance exceeded at.}
}
\value{
numeric value of distance in whatever units were used to create the
model.
}
\description{
\code{get_dist} returns the distance of the predator when a
specific \strong{{α}} or \strong{{dα/dt}} value is reached.
}
\details{
Input can be an object of class \code{attack_model},
\code{attack_model_whale}, or \code{data.frame} exported from either of
those functions. \strong{{α}} and \strong{{dα/dt}} columns are identified
automatically. Only one of the \emph{\code{dadt}} or \emph{\code{alpha}} inputs
should be entered.
}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
