% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{get_time}
\alias{get_time}
\title{extract attacker time to reach prey when a specific da/dt value is exceeded}
\usage{
get_time(mod, ALT)
}
\description{
extract attacker time to reach prey when a specific da/dt value is exceeded
}
\keyword{internal}
